connectd: handle custom messages.
This is neater than what we had before, and slightly more general.

Changelog-Changed: JSON_RPC: `sendcustommsg` now works with any connected peer, even when shutting down a channel.
