+++ connectd/connectd.c
@@ -1975,6 +1975,10 @@ static struct io_plan *recv_req(struct io_conn *conn,
 		onionmsg_req(daemon, msg);
 		goto out;
 
+	case WIRE_CONNECTD_CUSTOMMSG_OUT:
+		send_custommsg(daemon, msg);
+		goto out;
+
 	case WIRE_CONNECTD_DEV_MEMLEAK:
 #if DEVELOPER
 		dev_connect_memleak(daemon, msg);
@@ -1989,6 +1993,7 @@ static struct io_plan *recv_req(struct io_conn *conn,
 	case WIRE_CONNECTD_DEV_MEMLEAK_REPLY:
 	case WIRE_CONNECTD_PING_REPLY:
 	case WIRE_CONNECTD_GOT_ONIONMSG_TO_US:
+	case WIRE_CONNECTD_CUSTOMMSG_IN:
 		break;
 	}
