+++ lightningd/test/run-invoice-select-inchan.c
@@ -482,11 +482,6 @@ struct command_result *param_array(struct command *cmd UNNEEDED, const char *nam
 				   const char *buffer UNNEEDED, const jsmntok_t *tok UNNEEDED,
 				   const jsmntok_t **arr UNNEEDED)
 { fprintf(stderr, "param_array called!\n"); abort(); }
-/* Generated stub for param_bin_from_hex */
-struct command_result *param_bin_from_hex(struct command *cmd UNNEEDED, const char *name UNNEEDED,
-					  const char *buffer UNNEEDED, const jsmntok_t *tok UNNEEDED,
-					  u8 **bin UNNEEDED)
-{ fprintf(stderr, "param_bin_from_hex called!\n"); abort(); }
 /* Generated stub for param_bool */
 struct command_result *param_bool(struct command *cmd UNNEEDED, const char *name UNNEEDED,
 				  const char *buffer UNNEEDED, const jsmntok_t *tok UNNEEDED,
@@ -560,9 +555,6 @@ struct command_result *param_u64(struct command *cmd UNNEEDED, const char *name 
 /* Generated stub for peer_active_channel */
 struct channel *peer_active_channel(struct peer *peer UNNEEDED)
 { fprintf(stderr, "peer_active_channel called!\n"); abort(); }
-/* Generated stub for peer_get_owning_subd */
-struct subd *peer_get_owning_subd(struct peer *peer UNNEEDED)
-{ fprintf(stderr, "peer_get_owning_subd called!\n"); abort(); }
 /* Generated stub for peer_memleak_done */
 void peer_memleak_done(struct command *cmd UNNEEDED, struct subd *leaker UNNEEDED)
 { fprintf(stderr, "peer_memleak_done called!\n"); abort(); }
