--- wire/common_wire.csv
@@ -1,10 +1,0 @@
-# A custom message that we got from a peer and don't know how to handle, so we
-# forward it to the master for further handling.
-msgtype,custommsg_in,1030
-msgdata,custommsg_in,msg_len,u16,
-msgdata,custommsg_in,msg,u8,msg_len
-
-# A custom message that the master tells us to send to the peer.
-msgtype,custommsg_out,1031
-msgdata,custommsg_out,msg_len,u16,
-msgdata,custommsg_out,msg,u8,msg_len
